Move git tests files to git/ folder.
