+++ ide/app/test/all.dart
@@ -14,21 +14,22 @@ import 'analyzer_test.dart' as analyzer_test;
 import 'app_test.dart' as app_test;
 import 'compiler_test.dart' as compiler_test;
 import 'editors_test.dart' as editors_test;
-import 'file_operations_test.dart' as file_operations_test;
 import 'files_test.dart' as files_test;
-import 'git_object_test.dart' as git_object_test;
-import 'git_objectstore_test.dart' as git_objectstore_test;
-import 'git_pack_test.dart' as git_pack_test;
-import 'git_pack_index_test.dart' as git_pack_index_test;
-import 'git_test.dart' as git_test;
-import 'git_utils_test.dart' as git_utils_test;
+import 'git/file_operations_test.dart' as git_file_operations_test;
+import 'git/git_test.dart' as git_test;
+import 'git/object_test.dart' as git_object_test;
+import 'git/objectstore_test.dart' as git_objectstore_test;
+import 'git/pack_test.dart' as git_pack_test;
+import 'git/pack_index_test.dart' as git_pack_index_test;
+import 'git/zlib_test.dart' as zlib_test;
+import 'git/utils_test.dart' as git_utils_test;
 import 'preferences_test.dart' as preferences_test;
 import 'sdk_test.dart' as sdk_test;
 import 'server_test.dart' as server_test;
 import 'tcp_test.dart' as tcp_test;
 import 'utils_test.dart' as utils_test;
 import 'workspace_test.dart' as workspace_test;
-import 'zlib_test.dart' as zlib_test;
+
 
 /**
  * Place all new tests here.
@@ -42,8 +43,8 @@ void defineTests() {
   app_test.defineTests();
   compiler_test.defineTests();
   editors_test.defineTests();
-  file_operations_test.defineTests();
   files_test.defineTests();
+  git_file_operations_test.defineTests();
   git_object_test.defineTests();
   git_objectstore_test.defineTests();
   git_pack_test.defineTests();
