+++ ide/app/test/git/object_test.dart
@@ -6,7 +6,7 @@ library git_object_test;
 
 import 'package:unittest/unittest.dart';
 
-import '../lib/git/object.dart';
+import '../../lib/git/object.dart';
 
 defineTests() {
   group('git.objects', () {
