+++ ide/app/test/git/zlib_test.dart
@@ -10,7 +10,7 @@ import 'dart:typed_data';
 import 'package:unittest/unittest.dart';
 import 'package:utf/utf.dart';
 
-import '../lib/git/zlib.dart';
+import '../../lib/git/zlib.dart';
 
 final String ZLIB_INPUT_STRING = "This is a test.";
 final String ZLIB_DEFLATE_OUTPUT =
